Migrate BashTub Tests to TestContainers Based Tests
1. Since BashTub framework has not been supportted for the last 7 years
2. SInce having a Golang compiler protect us while running tests
3. Since TestContainers have been successfully used for a PoC
   to Use them in our Tests

The bashtub Tests have been migrated to the new TestContainers Suite,
which gives a better assurance since Tests are running in a local
docker based cluster.

Only BashTub Tests that have not been migrated are 2 tests related to
auth, which will be moved when the Auth project is started.

Linear Task: https://linear.app/expanso/issue/ENG-278/evaluate-bashtab-and-migrate-from-them-if-needed
Github Issue: #4629