Remove Depandabot Auto Merges and CircleCI References (#4660)
Currently We have multiple implementation of Dependabot, one of them did
not run for months, and the default one keeps failing. Also, Circle CI
is still being referenced for some reason in the Dependabot flow.

For this PR:
1. Remove auto-merge logic for Dependabot until we consolidate CI to
Github Actions. The auto-merge setup had many bugs, as well as duplicate
functionality.
2. For now, rely of the PRs created by Depandabot, and we will do manual
PR reviews and merges for the PRs created by Depandabot.
3. When all of the CI is consolidated to Github Actions, we can bring
back the automatic Depandabot merges, in a cleaner way.
4. Remove any CircleCI References. It is confusing to keep them around
for no reason.