integrate lib/watcher with job store (#4676)
The PR replaces the legacy watcher with the new lib/watcher library in
the job store to record execution upserts and evaluation events. No use
of the library's checkpointing capabilities yet and only watching
evaluation events from the tip for now similar to the legacy
implementation. That will happen in upcoming PRs.