Support NATS TLS Communication
This commit add the ability to enable TLS for NATS communication
in a cluster. It supports custom CA as well as regular signed
certificate by known CA.

To Enable TLS for the NATS Server on the orchestrator node and
compute node, please check the integration tests in this commit
as they have detailed coverage for multiple usecases as well as
detailed configuration files.

Linear: https://linear.app/expanso/issue/ENG-305/plan-and-implement-nats-tls-enabling-in-core-bacalhau
Github: #4681