Support NATS TLS Communication (#4686)
This PR adds the ability to enable TLS for NATS communication in a
cluster.
It supports custom CA as well as regular signed certificate by known CA.
If the TLS configuration was not set in the config, it will default back
to non-tls. The NATS server within the orchestrator node dictates if
NATS TLS is enabled or not. When enabled, only TLS communication is
allowed.

Sample Orchestrator Config:
```yaml
NameProvider: "uuid"
API:
  Port: 1234
Orchestrator:
  Enabled: true
  Auth:
    Token: "i_am_very_secret_token"
  TLS:
    ServerCert: "/path/to/cert"
    ServerKey: "/path/to/key"
    CACert: "/path/to/ca-cert" # Optional, will default to the system trust-store 
    ServerTimeout: 15
```

Sample Compute Node Config:
```yaml
NameProvider: "uuid"
API:
  Port: 1234
Compute:
  Enabled: true
  Orchestrators:
    - nats://bacalhau-orchestrator-node:4222
  Auth:
    Token: "i_am_very_secret_token"
  TLS:
    CACert: "/path/to/ca-cert" # Optional, will default to the system trust-store 

```

For details on how to enable TLS for the NATS Server on the orchestrator
node and compute node, please check the integration tests in the commits
within this PR as they have detailed coverage for multiple use-cases as
well as detailed configuration files.



Linear:
https://linear.app/expanso/issue/ENG-305/plan-and-implement-nats-tls-enabling-in-core-bacalhau
Github: #4681