Allow to pass URL as the API Host for CLI
When the CLI tries to connect to the orchestrator API over HTTP, the
the current configuration enforces the user to specify the PORT and if
they should connect over TLS or not.

In this change, the user can specify the Host as a URL (including
scheme and port), the right connection will be made.

This is a backward compatible change, the current behavior is still
valid. Though if the Host is a URL, it will take precedence over other
params (port and useTLS).

Example, now BACALHAU_API_HOST can be defined as:
'http://bacalhau-orchestrator-node:5555' or
'https://bacalhau-orchestrator-node:4444' or
'http://1.2.3.4:3333' or
'https://1.2.3.4' or
'https://1.2.3.4:7777' etc...

Linear: https://linear.app/expanso/issue/ENG-400/allow-api-host-to-accept-a-url