Enhance value column configuration with width constraints in config list (#4848)
Closes #4789 
This pull request includes a change to the `cmd/cli/config/list.go` file
to improve the display of the "Value" column in the configuration list
table.

Enhancements to table column configuration:

*
[`cmd/cli/config/list.go`](diffhunk://#diff-30602d5cf9bbe56af2982dd1b800c529ec64a62985243376475e16c75b89c971L102-R102):
Added `WidthMax` and `WidthMaxEnforcer` properties to the "Value" column
configuration to ensure the column width does not exceed 80 characters
and to wrap text softly if it does.

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **New Features**
- Enhanced the display of configuration information by enforcing a
maximum width of 80 characters and applying soft text wrapping. This
update improves the overall layout when displaying configuration data,
ensuring that long entries are neatly formatted. Users will experience a
clearer and more user-friendly presentation when reviewing configuration
values.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->