fix node update annotation & regenerate swagger (#4931)
The update node swagger annotation was listing the operation as a POST
request when the code shows it is used as a PUT request. This PR updates
the annotation (`node.go`) and regenerates the swagger schemas.

### Testing

1. `make build-dev` and test webui locally
2. `make test && make integration-test`


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Documentation**
- Improved API documentation for orchestrator node endpoints, clarifying
the usage of GET and PUT methods for retrieving and updating nodes by
ID.
- Added clearer descriptions, summaries, and error responses (including
404 Not Found) to the API specification and Swagger documentation.
- Updated API schema to better reflect standard RESTful practices,
separating retrieval and update operations.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->