Add automatic volume creation for local input source
- Change the spec value for local provider from 'localDirectory' to 'local'
- Introduce CreateAs param to handle non-existent volumes
- Support creation of both files and directories based on source path
- Add inference logic to determine appropriate creation strategy
- Improve error messaging

Currently if a local input source (file or directory) specified in the Job spec does not exist on a compute node, it will not try to bid for the execution. This changes the behavior to allow compute nodes to bid if they can create an empty directory or file at the specified path. When such a bid is accepted, the compute node will create an empty file or directory and mount it as normal. This change also introduces a new parameter to the Job spec input source 'CreateAs', which allows the user to specify how they want the input source to be created. By default the system will try to infer the volume type (file/directory) based on the provided path.